[DEV-1960] Refactoring
